Partial implementation of matrix multiplication with a scalar
`operator*=` tests have been run and run successfully, but the other
operators (two versions of `operator*` between matrix and scalar) don't
compile due to the fact that they are too ambiguous.

For this I can foresee two solutions (I don't have internet connection
right now), heavy use of template metaprogramming and `std::enable_if`
or "concepts".

I'd prefer the latter, but I'm not sure if any compiler actually
implements them already. If so, we are going to go with those, otherwise
we are going with the former.